clean code: move dynamic class to separate file
